requirements.txt
redlock
python-redis-lock
redis
retrying
